requirements.txt
apify<3.0.0
apify_client<2.0.0
crewai>=0.102.0
crewai-tools>=0.36.0
python-dotenv>=0.19.0
playwright>=1.41.0
html2text>=2020.1.16
crawlee==0.6.5